+++ SpatialScan/likelihood.py
