+++ SpatialScan/results.py
