+++ SpatialScan/timeseries.py
