+++ doc/timeseries.html
